add neural ode results
