+++ dl/assets/references/dl-fundamentals.bib
@@ -608,3 +608,20 @@ NOTE    = {revision \#138057}
   url    = "https://www.ai.rug.nl/minds/uploads/ESNTutorialRev.pdf",
   file   = "All Papers/JAEGER/Jaeger and Kirkpatrick - A tutorial on training recurrent neural networks, covering BPTT, RTRL, EKF and the 'echo state network' approach.pdf"
 }
+
+
+@software{Poli_TorchDyn_Implicit_Models,
+author = {Poli, Michael and Massaroli, Stefano and Yamashita, Atsushi and Asama, Hajime and Park, Jinkyoo and Ermon, Stefano},
+title = {{TorchDyn: Implicit Models and Neural Numerical Methods in PyTorch}},
+url = {https://github.com/DiffEqML/torchdyn}
+}
+
+@software{Chen_torchdiffeq_2021,
+author = {Chen, Ricky T. Q.},
+license = {MIT},
+month = jun,
+title = {{torchdiffeq}},
+url = {https://github.com/rtqichen/torchdiffeq},
+version = {0.2.2},
+year = {2021}
+}
